refactor: change id fields from int to long

+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -47,34 +47,32 @@ public class UserController {
 
     //получить пользователя по id
     @GetMapping("/{id}")
-    public User getUserById(@PathVariable @Min(1) int id) {
+    public User getUserById(@PathVariable @Min(1) long id) {
         return userStorage.getUserById(id);
     }
 
     //добавление в друзья
     @PutMapping("/users/{id}/friends/{friendId}")
-    public User addFriend(@PathVariable("id") @Min(1) int userId, @PathVariable @Min(1) int friendId) {
+    public User addFriend(@PathVariable("id") @Min(1) long userId, @PathVariable @Min(1) long friendId) {
         return userService.addFriend(userId, friendId);
     }
 
     //удаление из друзей
     @DeleteMapping("/{id}/friends/{friendId}")
-    public User removeFriend(@PathVariable("id") @Min(1) int userId, @PathVariable @Min(1) int friendId) {
+    public User removeFriend(@PathVariable("id") @Min(1) long userId, @PathVariable @Min(1) long friendId) {
         return userService.removeFriend(userId, friendId);
     }
 
     //возвращает список пользователей, являющихся его друзьями
     @GetMapping("/{id}/friends")
-    public Set<User> getFriends(@PathVariable @Min(1) int id) {
+    public Set<User> getFriends(@PathVariable @Min(1) long id) {
         return userStorage.getUserById(id).getFriends();
     }
 
     //список друзей, общих с другим пользователем
     @GetMapping("/users/{id}/friends/common/{otherId}")
-    public Set<User> getMutualFriends(@PathVariable("id") @Min(1) int user1Id
-            , @PathVariable("otherId") @Min(1) int user2Id) {
+    public Set<User> getMutualFriends(@PathVariable("id") @Min(1) long user1Id
+            , @PathVariable("otherId") @Min(1) long user2Id) {
         return userService.getMutualFriends(user1Id, user2Id);
     }
 }
-
-//todo логирование
